Regional builds should not display the map in triplicate
The map-triplicate functionality is useful for interpreting sequential transmissions which involve the entire world, however this comes at a performance cost (more elements in the DOM, and larger arrays for the map logic to loop through) as well as producing default map-bounds which nearly always show the entire world (although the default views are by no means perfect with the changes in this PR).

For regional nCoV builds, the focus is on an individual location and thus it's not necessary to display the map in triplicate.